A-Duy small update

+++ src/main/java/vn/codegym/flightagency/dto/PassengerInfoDTO.java
@@ -19,5 +19,5 @@ public class PassengerInfoDTO {
 
     private String gender;
 
-    private Double baggagePrice;
+    private int baggagePrice;
 }
